adjust value width and better config eval

+++ src/aircraft/Aircraft.hpp
@@ -65,8 +65,8 @@ public:
      * @param head    The heading
      */
     Aircraft(std::string& r_id, struct util::GpsPosition& r_pos, double gnd_spd,
-            std::uint32_t id_t, std::int32_t ac_t, double climb_r, /*double turn_r,*/
-            double head);
+             std::uint32_t id_t, std::int32_t ac_t, double climb_r, /*double turn_r,*/
+             double head);
     /**
      * @fn ~Aircraft
      * @brief Destructor
@@ -149,7 +149,7 @@ public:
      * @brief Get the current update age.
      * @return the update age
      */
-    inline std::uint32_t& getUpdateAge()
+    inline std::uint64_t& getUpdateAge()
     {
         return mUpdateAge;
     }
@@ -305,8 +305,9 @@ public:
         mFullInfo = info;
     }
 
-    inline std::uint32_t& getUpdateAttempts() {
-return mUpdateAttempts;
+    inline std::uint64_t& getUpdateAttempts()
+    {
+        return mUpdateAttempts;
     }
 
 private:
@@ -321,11 +322,11 @@ private:
     /// Is full set of information available?
     bool mFullInfo = false;
     /// Times processed without update.
-    std::uint32_t mUpdateAge = 0;
+    std::uint64_t mUpdateAge = 0;
     /// Got last update with which priority.
     std::uint32_t mLastPriority = 0;
     ///
-    std::uint32_t mUpdateAttempts = 0;
+    std::uint64_t mUpdateAttempts = 0;
     /// Last registered position.
     struct util::GpsPosition mPosition;
     /// Speed over ground; m/s
