+++ src/data/Data.hpp
@@ -59,7 +59,7 @@ public:
 	 * @tparam T The new data
 	 * @param prio The priority of the update
 	 */
-	virtual void update(const T&, std::uint32_t vPriority, std::uint32_t& rAttempts) = 0;
+	virtual void update(const T&, std::uint32_t vPriority, std::uint64_t& rAttempts) = 0;
 
 	/**
 	 * @fn init
